chore(common): remove base-alloy-rpc-jsonrpsee crate
Move MinerApiExt trait into base-execution-rpc where its server
implementation already lives. Inline a client-only MinerApiExt
in the batcher to avoid pulling reth deps. Move EthSignerApi into
base-alloy-signer, its sole consumer. Delete the now-empty
base-alloy-rpc-jsonrpsee crate along with the unused BaseAdminApi
trait (superseded by base-consensus-rpc's AdminApi).
