--- crates/common/rpc-jsonrpsee/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "base-common-rpc-jsonrpsee"
-description = "Base RPC Client"
-
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-exclude.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# Alloy
-alloy-primitives = { workspace = true, features = ["serde"] }
-alloy-rpc-types-eth = { workspace = true, features = ["std", "serde"], optional = true }
-
-# rpc
-jsonrpsee = { workspace = true, features = ["macros", "server"] }
-
-[features]
-client = [ "jsonrpsee/async-client", "jsonrpsee/client" ]
-signer = [ "dep:alloy-rpc-types-eth" ]
