+++ crates/execution/node/src/node.rs
@@ -6,8 +6,8 @@ use alloy_consensus::BlockHeader;
 use alloy_primitives::{Address, B64, B256};
 use base_common_chains::BaseUpgrades;
 use base_common_consensus::BasePrimitives;
-use base_common_rpc_jsonrpsee::MinerApiExtServer;
 use base_common_rpc_types_engine::{BasePayloadAttributes, ExecutionData};
+use base_execution_rpc::MinerApiExtServer;
 use base_execution_chainspec::BaseChainSpec;
 use base_execution_consensus::OpBeaconConsensus;
 use base_execution_evm::{BaseEvmConfig, OpRethReceiptBuilder};
