+++ crates/execution/rpc/Cargo.toml
@@ -56,7 +56,6 @@ alloy-rpc-types-eth.workspace = true
 alloy-transport-http.workspace = true
 alloy-rpc-types-debug.workspace = true
 alloy-rpc-types-engine.workspace = true
-base-common-rpc-jsonrpsee.workspace = true
 base-common-network = { workspace = true, features = ["reth"] }
 base-common-consensus = { workspace = true, features = ["reth"] }
 base-common-rpc-types = { workspace = true, features = ["reth"] }
@@ -69,7 +68,7 @@ futures.workspace = true
 async-trait.workspace = true
 
 # rpc
-jsonrpsee.workspace = true
+jsonrpsee = { workspace = true, features = ["macros", "server"] }
 jsonrpsee-core.workspace = true
 jsonrpsee-types.workspace = true
