+++ crates/execution/runner/src/add_ons.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use base_common_rpc_jsonrpsee::MinerApiExtServer;
+use base_execution_rpc::MinerApiExtServer;
 use base_execution_payload_builder::{
     Attributes, PayloadPrimitives,
     config::{GasLimitConfig, OpDAConfig},
